requirements.txt
fastapi
uvicorn
httpx
bleak
python-dotenv
